Faster and smarter dependencies management
Refactored the code to handle cross-stacks dependencies more deeply and
enable parallelizing the operations on the stacks (deploying, deleting
etc.). The goal is to speed up the operations when working with a long
list of inter-dependent stacks.

This was done with minor adjustements to the previous API. The main
difference is the introduction of a dependsOn argument to list a stack's
dependent stacks, and a breaking change in how dynamic stack parameters
are resolved. The new logic uses stack names and regions to retrieve
their outputs dynamically, instead of passing those along in an ordered
list during the sequential processing of each stack.

The tracker outputs were simplified as a result since multiple
operations happen at the same time, but key information remains.