Increased default credentials provider's timeout
Previously, intermittent errors were raised from the credentials
provider which appeared to be due to timeout happenning on EC2 as the
corresponding credentials provider timed out while retrieving the
instance's role from its metadata. In this fix, the default credentials
provider used by the various AWS SDK clients was specified with an
increased timeout of 5s instead of the default 1s.