Pruning optimization (#7)
* Improved the deployment files pruning process

Revamped the pruning process of old deployment files to be much faster
than before. Previously, many checks were made on each file throughout
the process, including getting a file's tag which can only be done on a
per-file basis and ran particularly slow with the throttle.
The new approach optimizes the process so that files are only checked
once, and less files are required to be checked.

Fixed several bugs related to the deployment files tagging.

Implemented StackParams and ResolvedStackParams as actual classes,
rather than typing hints.